[REFACTOR]custom hook 'useScrollNextSection' 추가
